% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{cellfreetranscriptome}
\alias{cellfreetranscriptome}
\alias{cellfreetranscriptome-package}
\title{A comprehensive characterization of the cell-free transcriptome reveals tissue- and subtype-specific biomarkers for cancer detection}
\description{
Provides code base for reproducing results of the manuscript.
}
\author{
Ruth Mauntz (rmauntz@grail.com)
}
