% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_id.R
\name{convert_gene_id}
\alias{convert_gene_id}
\title{Convert the gene_id to gene_symbols}
\usage{
convert_gene_id(data)
}
\arguments{
\item{data}{a data frame for the expression matrix. The first column is the
feature name and each column corresponds to a sample.}
}
\value{
the data whose first column is converted to gene symbol.
}
\description{
Convert the gene_id to gene_symbols
}
