% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{count_normalization}
\alias{count_normalization}
\title{RNA-seq reads normalization using size factor}
\usage{
count_normalization(count_raw, sample_anno)
}
\arguments{
\item{count_raw}{: The numeric matrix of raw reads count.}

\item{sample_anno}{: The metadata of the samples.}
}
\value{
The normalized reads count matrix.
}
\description{
The definition of size factor is in this paper:
\url{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2010-11-10-r106}
}
