% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_two_group_de_te.R
\name{edger_two_group_de_te}
\alias{edger_two_group_de_te}
\title{Find differentiated expressed genes in two classes of RNA-seq data}
\usage{
edger_two_group_de_te(genetable, labels, controls, method = "TMMwsp")
}
\arguments{
\item{genetable}{a data_frame with columns gene_id and columns corresponding
to the numeric value of RNA counts per gene in a sample.}

\item{labels}{a vector of length ncol(genetable)-1, corresponding to the sample groups}

\item{controls}{vector of genes that you expect to have consistent expression between groups}

\item{method}{character, normalization method used by edgeR}
}
\value{
data frame with differential p-values for each gene
}
\description{
Takes in a vector of control / housekeeping genes to use for expression
analysis of a targeted RNA panel. edgeR normalization assumes that most of
the genes are not differentially expressed, which might not be true in a
targeted panel (https://support.bioconductor.org/p/87121/).
}
