% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{genetable_to_matrix}
\alias{genetable_to_matrix}
\title{Convert a gene expression data table into a caret-friendly matrix}
\usage{
genetable_to_matrix(expression, transpose = TRUE)
}
\arguments{
\item{expression}{A table with a column "gene_id" that contains expression estimates.}

\item{transpose}{Set to \code{TRUE} to transpose rows and columns.}
}
\value{
A matrix where genes are columns and samples are rows.
}
\description{
Gene expression data tables follow tidy conventions where a separate gene_id column is used in
place of row names. Caret expects explanatory variables (in this case genes) to be columns.
This function transposes the rna expression table and adds the gene_ids as column names.
}
