% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_darkness_index.R
\name{get_darkness_index}
\alias{get_darkness_index}
\title{Get gene darkness index from pilot healthy}
\usage{
get_darkness_index(spliced_count_file)
}
\arguments{
\item{spliced_count_file}{a data frame for the expression matrix. The first column is the
feature name and each column corresponds to a sample.}
}
\value{
data frame containing the median and standard deviation expression values for each gene
}
\description{
Get gene darkness index from pilot healthy
}
