% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{get_rpm}
\alias{get_rpm}
\title{Normalize the raw count data by size factor}
\usage{
get_rpm(genetable)
}
\arguments{
\item{genetable}{A data frame for the expression matrix. The first column is the
feature name and each column corresponds to a sample.}
}
\value{
A data frame with rpm data.
}
\description{
Normalize the raw count data by size factor
}
