% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Create a standard hue of colors for ggplot2}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{integer, number of colors to return for plotting.}
}
\value{
character, a hexcode set of colors of length n.
}
\description{
Create a standard hue of colors for ggplot2
}
