% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{matrix_to_genetable}
\alias{matrix_to_genetable}
\title{Convert a caret-friendly matrix into a gene expression data table}
\usage{
matrix_to_genetable(expression_matrix, genes_as_cols = TRUE)
}
\arguments{
\item{expression_matrix}{A matrix where column names correspond to gene_id.}

\item{genes_as_cols}{In classical ML, features are columns. Sometimes we have matrices where
features (genes) are rows. Set this to \code{FALSE} in order to handle this case.}
}
\value{
A data table with gene_id as a column.
}
\description{
This function transposes the rna expression table and adds the gene_ids column.
Gene expression data tables follow tidy conventions where a separate gene_id column is used in
place of row names. Caret expects explanatory variables (in this case genes) to be columns.
}
