% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize reference matrix for tissue deconvolution}
\usage{
normalize_matrix(signatures)
}
\arguments{
\item{signatures}{a data_frame reference matrix}
}
\value{
normalized reference matrix
}
\description{
Normalize reference matrix for tissue deconvolution
}
