% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_sig_level}
\alias{set_sig_level}
\title{Translate p-values from significance testing to astericks}
\usage{
set_sig_level(p.value)
}
\arguments{
\item{p.value}{numeric, a p-value}
}
\value{
character, an asterick associated with the given
p-value. < 0.001 yields `***`, < 0.01 yields `**``,
 < 0.05 yields `*`, and < 0.001 yields ``.
}
\description{
Translate p-values from significance testing to astericks
}
