% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_rpms.R
\name{summarise_rpms}
\alias{summarise_rpms}
\title{Summarises gene-wise RPMs across a vector of samples}
\usage{
summarise_rpms(gene_ids, rpms, th_rpm = 0)
}
\arguments{
\item{gene_ids}{vector of gene ids}

\item{rpms}{matrix of gene-wise rpms, number of rows are length(gene_ids)}

\item{th_rpm}{numeric threshold RPM limit to binarise gene as detected (1) or not detected (0)}
}
\value{
A data_frame with one row per gene id, columns containing number detected above th_rpm,
sum_rpms by gene, mean_rpm by gene, and sd_rpm (standard deviation) by gene
}
\description{
Summarises gene-wise RPMs across a vector of samples
}
