% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueDeconResidual.R
\name{tissueDeconResidual}
\alias{tissueDeconResidual}
\title{Estimates tissue deconvolution using quadprog}
\usage{
tissueDeconResidual(signatures, dataset_DeconRNASeq)
}
\arguments{
\item{signatures}{data_frame with columns containing tissue-specific gene of different tissue.}

\item{dataset_DeconRNASeq}{data_frame with columns of mixture RNA-seq samples.}
}
\value{
A data_frame with one row per containing the estimation of tissue fraction
}
\description{
Estimates tissue deconvolution using quadprog
}
