% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test.R
\name{wilcoxon_test}
\alias{wilcoxon_test}
\title{Non-parametric Wilcoxon significance test between groups}
\usage{
wilcoxon_test(rpm_df, gene, group1, group2)
}
\arguments{
\item{rpm_df}{dataframe with columns Subtype, gene_id,
and expression_new}

\item{gene}{character gene name}

\item{group1}{first group in dataframe to test}

\item{group2}{second group in dataframe to test}
}
\value{
a numeric p-value
}
\description{
Non-parametric Wilcoxon significance test between groups
}
